Change the Nixpkgs version to v<YY><MM>.<count>.0
We don't want to encode the month as the minor version since that
implies the November release (23.11) is backwards compatible with the
May release (23.5). So better to encode it as v2305 and v2311.

The revcount should be stored as the minor version rather than the
patch version, since commits to stable branches *can* add new
functionality.